fixup! mingw: allow `git.exe` to be used instead of the "Git wrapper"
The original test case obviously only worked in the x86_64 flavor of Git
for Windows (and since I rarely test on i686, this was not caught).

This completely breaks on Windows/ARM64:

  +++ GIT_DIR=/dev/null
  +++ git diff --no-index --ignore-cr-at-eol -- expect actual
  diff --git a/expect b/actual
  index e0dc09e..2301bbf 100644
  --- a/expect
  +++ b/actual
  @@ -1,3 +1,3 @@
  -MSYSTEM=CLANGARM64
  +MSYSTEM=MINGW64
   mingw64
   usr
  error: last command exited with $?=1
  not ok 221 - MSYSTEM/PATH is adjusted if necessary

To fix that, let's rely on `MINGW_PREFIX` (falling back to constructing
it from `MSYSTEM`, defaulting to `MINGW64` if _that_ is unset, too).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>